Fix bootstrap script where circular symlinks were created
If a symlink was trying to create a link between two directories, and
the destination directory already existed, ln would create a link inside
the existing destination directory when using the -f option. To avoid circular
links, we check if the source is a directory first, and if the
destination directory already exists, we skip creating a symlink for it.